+++ duilib/Image/Image_WEBP.cpp
@@ -216,7 +216,7 @@ bool Image_WEBP::IsDelayDecodeFinished() const
     if (m_impl->m_bAsyncDecoding) {
         return false;
     }
-    return (m_impl->m_frames.size() + m_impl->m_delayFrames.size()) == m_impl->m_nFrameCount;
+    return (int32_t)(m_impl->m_frames.size() + m_impl->m_delayFrames.size()) == m_impl->m_nFrameCount;
 }
 
 uint32_t Image_WEBP::GetDecodedFrameIndex() const
@@ -243,7 +243,7 @@ bool Image_WEBP::DelayDecode(uint32_t nMinFrameIndex, std::function<bool(void)> 
     uint32_t nFrameIndex = (uint32_t)(m_impl->m_delayFrames.size() + m_impl->m_frames.size());
     while (((IsAborted == nullptr) || !IsAborted()) &&
            (nMinFrameIndex >= (int32_t)(m_impl->m_frames.size() + m_impl->m_delayFrames.size())) &&
-           ((m_impl->m_frames.size() + m_impl->m_delayFrames.size()) < m_impl->m_nFrameCount)) {
+           ((int32_t)(m_impl->m_frames.size() + m_impl->m_delayFrames.size()) < m_impl->m_nFrameCount)) {
         AnimationFramePtr pNewAnimationFrame;
         pNewAnimationFrame = DecodeImage_WEBP(m_impl->m_pWebPAnimDecoder,
                                               fImageSizeScale,
@@ -276,7 +276,7 @@ bool Image_WEBP::MergeDelayDecodeData()
     }
     if (!m_impl->m_bAsyncDecoding) {
         //如果解码完成，则释放图片资源
-        if (m_impl->m_frames.size() == m_impl->m_nFrameCount) {
+        if ((int32_t)m_impl->m_frames.size() == m_impl->m_nFrameCount) {
             if (m_impl->m_pWebPAnimDecoder != nullptr) {
                 WebPAnimDecoderDelete(m_impl->m_pWebPAnimDecoder);
                 m_impl->m_pWebPAnimDecoder = nullptr;
@@ -366,7 +366,7 @@ bool Image_WEBP::ReadFrameData(int32_t nFrameIndex, const UiSize& /*szDestRectSi
     if (!m_impl->m_bAsyncDecode) {
         //同步解码的情况, 解码所需要的帧
         while ((nFrameIndex >= (int32_t)m_impl->m_frames.size()) &&
-               (m_impl->m_frames.size() < m_impl->m_nFrameCount)) {
+               ((int32_t)m_impl->m_frames.size() < m_impl->m_nFrameCount)) {
             ASSERT(m_impl->m_delayFrames.empty());
             uint32_t nInitFrameIndex = (uint32_t)m_impl->m_frames.size();
             float fImageSizeScale = m_impl->m_fImageSizeScale;
